fix: api client should not be a singleton
